backend/requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
httpx==0.28.1
chromadb==1.0.15
langchain_huggingface==0.3.0
langchain_chroma==0.2.4
sentence-transformers==5.0.0
pinecone
python-dotenv==1.1.1